Removed calls to ensurePath
Zookeeper operations must fail if the ZK cluster ID
is missing or wrong. This makes sure agents only talk
to ZK clusters they are intended to. With this model
we cannot call ensurePath, because that would create
the expected path. Instead we must create individual
nodes under an existing path.